Move debezium mysql integration test to its own file
